web_agent_app/requirements.txt
# FastAPI和Web服务器
fastapi==0.100.1
uvicorn[standard]==0.23.2
python-multipart==0.0.6

# 现有系统依赖
openai==1.3.7
requests==2.31.0
python-dotenv==1.0.0

# 数据处理
pydantic==2.4.2

# 可选：生产环境和开发工具
# gunicorn==21.2.0  # 生产环境WSGI服务器
# pytest==7.4.3    # 测试框架
# httpx==0.25.2     # 用于测试API
